cleaned up chat object
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@7059 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ engine/src/game.cpp
@@ -771,7 +771,7 @@ void IGame::onTick(const float dt) {
 flip:
 		float fr = Window->getFrameRate();
 		if (_show_fps) {
-			_fps->hp = floor(fr);
+			_fps->hp = (int)floor(fr);
 			_fps->render(window, window.getWidth() - (int)(_fps->size.x * 3), window.getHeight() - (int)_fps->size.y);
 		}
 		if (_show_log_lines) {
